カスタムユニットダイアログのヒントテキストを改善し、アプリのルーター設定を更新。メニューページを削除し、AppPageに置き換え。
